+++ libs/coin-modules/coin-module-boilerplate/src/api/index.ts
@@ -27,21 +27,13 @@ export function createApi(config: BoilerplateConfig): Api<BoilerplateToken> {
   };
 }
 
-async function craft(
-  address: string,
-  transaction: {
-    recipient: string;
-    amount: bigint;
-    fee: bigint;
-  },
-): Promise<string> {
-  const nextSequenceNumber = await getNextValidSequence(address);
+async function craft(transactionIntent: TransactionIntent<BoilerplateToken>): Promise<string> {
+  const nextSequenceNumber = await getNextValidSequence(transactionIntent.sender);
   const tx = await craftTransaction(
-    { address, nextSequenceNumber },
+    { address: transactionIntent.sender, nextSequenceNumber },
     {
-      recipient: transaction.recipient,
-      amount: new BigNumber(transaction.amount.toString()),
-      fee: new BigNumber(transaction.fee.toString()),
+      recipient: transactionIntent.recipient,
+      amount: new BigNumber(transactionIntent.amount.toString()),
     },
   );
   return tx.serializedTransaction;