+++ libs/coin-modules/coin-tron/src/api/index.test.ts
@@ -9,12 +9,7 @@ import {
 } from "../logic";
 import coinConfig from "../config";
 import { TronConfig } from "../config";
-import {
-  Api,
-  TransactionIntent,
-  Pagination,
-  Transaction,
-} from "@ledgerhq/coin-framework/api/types";
+import { Api, Pagination, TransactionIntent } from "@ledgerhq/coin-framework/api/types";
 import { createApi } from ".";
 import { TronToken } from "../types";
 
@@ -68,21 +63,21 @@ describe("createApi", () => {
 
   it("should pass parameters well", async () => {
     const api: Api<TronToken> = createApi(mockTronConfig);
-
-    // Simulate calling all methods
-    await api.broadcast("transaction");
-    api.combine("tx", "signature", "pubkey");
-    await api.craftTransaction("address", {} as Transaction, "pubkey");
-    await api.estimateFees({
+    const intent: TransactionIntent<TronToken> = {
       type: "send",
-      sender: "address",
-      recipient: "address",
+      sender: "sender",
+      recipient: "recipient",
       amount: BigInt(10),
       asset: {
         standard: "trc10",
         tokenId: "1002000",
       },
-    } satisfies TransactionIntent<TronToken>);
+    };
+    // Simulate calling all methods
+    await api.broadcast("transaction");
+    api.combine("tx", "signature", "pubkey");
+    await api.craftTransaction(intent);
+    await api.estimateFees(intent);
     await api.getBalance("address");
     await api.lastBlock();
     await api.listOperations("address", {} as Pagination);
@@ -90,17 +85,8 @@ describe("createApi", () => {
     // Test that each of the methods was called with correct arguments
     expect(broadcast).toHaveBeenCalledWith("transaction");
     expect(combine).toHaveBeenCalledWith("tx", "signature", "pubkey");
-    expect(craftTransaction).toHaveBeenCalledWith("address", {}, "pubkey");
-    expect(estimateFees).toHaveBeenCalledWith({
-      type: "send",
-      sender: "address",
-      recipient: "address",
-      amount: BigInt(10),
-      asset: {
-        standard: "trc10",
-        tokenId: "1002000",
-      },
-    } satisfies TransactionIntent<TronToken>);
+    expect(estimateFees).toHaveBeenCalledWith(intent);
+    expect(craftTransaction).toHaveBeenCalledWith(intent);
     expect(getBalance).toHaveBeenCalledWith("address");
     expect(lastBlock).toHaveBeenCalled();
     expect(listOperations).toHaveBeenCalledWith("address", {});