+++ libs/device-react/CHANGELOG.md
@@ -1,11 +1,18 @@
 # @ledgerhq/device-react
 
-## 0.2.23-nightly.0
+## 0.2.23
 
 ### Patch Changes
 
 - Updated dependencies []:
-  - @ledgerhq/device-core@0.4.17-nightly.0
+  - @ledgerhq/device-core@0.4.17
+
+## 0.2.23-next.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @ledgerhq/device-core@0.4.17-next.0
 
 ## 0.2.22