+++ libs/ledgerjs/packages/swift-bridge-hw-app-eth/CHANGELOG.md
@@ -1,11 +1,18 @@
 # @ledgerhq/swift-bridge-hw-app-eth
 
-## 1.2.24-nightly.0
+## 1.2.24
 
 ### Patch Changes
 
 - Updated dependencies [[`9c4c82e`](https://github.com/LedgerHQ/ledger-live/commit/9c4c82e8e81aaadfc1132f0a25812be094e54453)]:
-  - @ledgerhq/hw-app-eth@6.43.0-nightly.0
+  - @ledgerhq/hw-app-eth@6.43.0
+
+## 1.2.24-next.0
+
+### Patch Changes
+
+- Updated dependencies [[`9c4c82e`](https://github.com/LedgerHQ/ledger-live/commit/9c4c82e8e81aaadfc1132f0a25812be094e54453)]:
+  - @ledgerhq/hw-app-eth@6.43.0-next.0
 
 ## 1.2.23